test: add test case auth & tls + fix: log format & appdir
