+++ examples/config/logging.toml
@@ -2,7 +2,7 @@ version_requirement = "<1.0.0"
 
 [server.logs]
 filepath = "/var/log/vsmtp/vsmtp.log"
-format = "{d} {l} - "
+format = "{d(%Y-%m-%d %H:%M:%S)} {h({l:<5} {I})} ((line:{L:<3})) $ {m}{n}"
 
 [server.logs.level]
 default = "warn"
