+++ vsmtp-config/src/tests/root_example/logging.rs
@@ -15,7 +15,7 @@ fn parse() {
             .with_ipv4_localhost()
             .with_logs_settings(
                 "/var/log/vsmtp/vsmtp.log",
-                "{d} {l} - ",
+                "{d(%Y-%m-%d %H:%M:%S)} {h({l:<5} {I})} ((line:{L:<3})) $ {m}{n}",
                 collection! {
                     "default".to_string() => log::LevelFilter::Warn,
                     "receiver".to_string() => log::LevelFilter::Info,
