+++ vsmtp-test/src/tests/mod.rs
@@ -14,40 +14,8 @@
  * this program. If not, see https://www.gnu.org/licenses/.
  *
 **/
-
-const TEST_SERVER_CERT: &str = "./src/tests/certs/certificate.crt";
-const TEST_SERVER_KEY: &str = "./src/tests/certs/privateKey.key";
-
 mod auth;
 mod clair;
 mod rset;
-mod starttls;
-mod tls_tunneled;
+mod tls;
 mod utf8;
-
-use vsmtp_config::Config;
-
-pub fn get_tls_config() -> Config {
-    Config::builder()
-        .with_version_str("<1.0.0")
-        .unwrap()
-        .with_server_name("testserver.com")
-        .with_user_group_and_default_system("root", "root")
-        .unwrap()
-        .with_ipv4_localhost()
-        .with_default_logs_settings()
-        .with_spool_dir_and_default_queues("./tmp/delivery")
-        .with_safe_tls_config(TEST_SERVER_CERT, TEST_SERVER_KEY)
-        .unwrap()
-        .with_default_smtp_options()
-        .with_default_smtp_error_handler()
-        .with_default_smtp_codes()
-        .without_auth()
-        .with_default_app()
-        .with_vsl("./src/empty_main.vsl")
-        .with_default_app_logs()
-        .without_services()
-        .with_system_dns()
-        .validate()
-        .unwrap()
-}
